install new icon packages
