+++ src/components/App.css
@@ -124,7 +124,7 @@
    flex-direction: column;
   
 }
-.navbar{
+.cart{
     width: 100%;
     height: 80px;
     background-color:#748D89;
