--- src/components/Navbar.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-import "./App.css"
-import { Link } from "react-router-dom";
-import{ShoppingCart} from "phosphor-react"
-
-export const Navbar = () => {
-    
-return (
- 
-      <div className="navbar">
-      <div className="links">
-      <Link to="/"> Shop </Link>
-      <Link to="/cart"> <ShoppingCart size={32}/>
-       </Link>
-
-      </div>
-   </div>
-
-)
-}
